refactor(market): a bit more of clean up
