+++ libs/solana-dapp/market/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["../../../../.eslintrc.json"],
+  "extends": ["../../../libs/solana-dapp/.eslintrc.json"],
   "ignorePatterns": ["!**/*"],
   "overrides": [
     {
