+++ libs/solana-dapp/market/README.md
@@ -1,0 +1,7 @@
+# solana-dapp-market
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test solana-dapp-market` to execute the unit tests via [Jest](https://jestjs.io).
