+++ libs/solana-dapp/market/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/get-prices';
 export * from './lib/get-price';
 export * from './lib/get-prices-from-wallet';
+export * from './lib/types';
