--- libs/solana-dapp/market/types/README.md
@@ -1,7 +1,0 @@
-# solana-dapp-market-types
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test solana-dapp-market-types` to execute the unit tests via [Jest](https://jestjs.io).
