--- libs/solana-dapp/market/utils/get-prices/README.md
@@ -1,7 +1,0 @@
-# solana-dapp-market-utils-get-prices
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test solana-dapp-market-utils-get-prices` to execute the unit tests via [Jest](https://jestjs.io).
