--- libs/solana-dapp/market/utils/get-prices/jest.config.js
@@ -1,15 +1,0 @@
-module.exports = {
-  displayName: 'solana-dapp-market-utils-get-prices',
-  preset: '../../../../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  transform: {
-    '^.+\\.[tj]sx?$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory:
-    '../../../../../coverage/libs/solana-dapp/market/utils/get-prices',
-};
