--- libs/solana-dapp/market/utils/get-tokens/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../../../../tsconfig.base.json",
-  "files": [],
-  "include": [],
-  "references": [
-    {
-      "path": "./tsconfig.lib.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
-    }
-  ]
-}
