+++ libs/solana-dapp/utils/get-tokens/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["../../../../../.eslintrc.json"],
+  "extends": ["../../../../libs/.eslintrc.json"],
   "ignorePatterns": ["!**/*"],
   "overrides": [
     {
