+++ libs/solana-dapp/utils/get-tokens/README.md
@@ -1,0 +1,7 @@
+# solana-dapp-utils-get-tokens
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test solana-dapp-utils-get-tokens` to execute the unit tests via [Jest](https://jestjs.io).
