+++ nx.json
@@ -63,13 +63,10 @@
     "solana-dapp-connection-types": {
       "tags": []
     },
-    "solana-dapp-market-types": {
+    "solana-dapp-market": {
       "tags": []
     },
-    "solana-dapp-market-utils-get-prices": {
-      "tags": []
-    },
-    "solana-dapp-market-utils-get-tokens": {
+    "solana-dapp-utils-get-tokens": {
       "tags": []
     },
     "solana-dapp-utils-u64": {
